requirements.txt
streamlit
pandas
undetected-chromedriver
selenium
Pillow
retrying
setuptools
xlsxwriter

